CASE-34 You can now be in a team.
AdminController probably the wrong place for this to live, and /admin
might not be appropriate URL at all for team stuff, but easy to move around.

If you are in a webgroup in-case-management-team-mentalhealth or similar
you will see additional links in the primary navigation. There's a conf key
you can override to use your own different set of webgroups, which will be recommended
particularly once the webgroups are being used in production.