Improve test application helpers.
Minimal Application wasn't actually minimal because it was still loading the default routes, and in turn
everything that any controller depended on. Now it has empty routes by default.

Added FakeExternalServer which makes it easier to run a fake external server based on some routes. Useful
for emulating Tabula or something. It runs as a separate Application of its own, which will ususally be minimal.