Server-Timing header.
Open to this being nixed as it's fairly intrusive on app code, requiring a TimingContext to be passed implicitly to wherever timing is required. I've tried to minimise the impact by making RequestContext and AuditLogContext implement TimingContext, so that a lot of the time the context is already available.

At the moment the only timing categories are db and tabula.