CASE-159 Fully verify type coming out of cache.
You may notice that there is a new dependency to handle this. I tried to do this with plain Scala but there's just no type information coming out of memcached - the only way to verify what's in the ServiceResult is to iterate the individual object and check that they are one, which is what this typeclass does. There might be some effect on performance since e.g. if the type is a list of X, it does check every item to make sure it's an X.