CASE-277 Fix source maps (for development).
If source maps isn't enabled for every plugin in the pipeline, they'll get lost or broken, so fix was to enable them on the plugins where it wasn't already enabled, and it also didn't seem to like working with inline source maps so changed to a separate file.

It still doesn't work with production which I am pretty sure is because of one of the minification steps we do that either doesn't have source maps enabled, or doesn't support it or has broken support.