CASE-448 Babelise stuff in node_modules.
Switches from .babelrc (file-relative config) to babel.config.js (project-relative config), which is the documentation-recommended way to be able to compile stuff in node_modules.

This change means that _all_ scripts we import from node_modules are run through Babel. This might not be what we want, in which case we can always add some whitelists to our babel-loader config to only compile the modules we want (wildcards might be smart enough to let us specify that anything in @universityofwarwick should be compiled)