Set Cross-Origin-Resource-Policy on user content
CORP is related to CORB. CORP is an opt-in mechanism which
stops no-cors requests from a different origin from succeeding. In other
words, this will prevent e.g. <img>, <script> and <object> embeds
of attachments within case management from a different origin.

Note that this is stronger than the protection offered by SameSite=Lax,
since it applies per-origin, not per-site.

This is implemented in Chrome and Safari, and set to be implemented
in Firefox too: https://www.chromestatus.com/feature/4647328103268352

Spec:

https://fetch.spec.whatwg.org/#cross-origin-resource-policy-header

Risks:

* There is a risk of breakage if we currently have web applications on a
  different domain != wellbeing-*.warwick.ac.uk which use the case
  management attachment system as a CDN for JavaScript and image
  resources. This seems unlikely.